se modifico el atributo stockquantity de integer a bigdecimal.

+++ src/main/java/com/japs/backend/supermercado/pos/domain/model/ProductSimple.java
@@ -20,7 +20,7 @@ public class ProductSimple {
 
     private BigDecimal price;
 
-    private Integer stockQuantity;
+    private BigDecimal stockQuantity;
 
     private String unitMeasurement;
